VFB-171: Place getCurrentYear and getCurrentMonth into common folder
