+++ src/common/date.ts
@@ -1,0 +1,9 @@
+export const getCurrentYear = (): number => {
+    const currentDate = new Date();
+    return currentDate.getFullYear();
+};
+
+export const getCurrentMonth = (): number => {
+    const currentDate = new Date();
+    return currentDate.getMonth() + 1;
+};
